Expanded UIElement functionality
It now has both a 'default action' and a 'custom action', plus a flag to determine which shall be used as the default. Default action accepts and returns a longlong int, a decent amount of data but nothing massive such that having one of these 'data blocks' with every UI element would use a crippling amount of space like using a 64 byte block might, as was being done before. Said block was replaced with a block pointer such that, should it be needed, the memory can be allocated to that specific UI element and the rest are left well enough alone.

Also, this means that a UI element can have no function at all.