Font system progress
Not functioning at the moment, but a quick rundown is that there are now 'unfinished objects' that just represent the full data needed to create an object. They also can be appended to each other. Plan for text is to create one of these unfinished objects and this can either be converted into an actual object, ui element, or appended onto a base UI element's appearance bits to add text to a button, for example.