More characters & Bugfix
More of the alphabet is complete and a bug was fixed with the spacing of the font. It is now properly consistent and is 'within' the font spacing system. As in, it abides by the rules of the font sizes and always has a relative spacing of 16 pixels vertically and 12 horizontally.