'Look at' works now- TODOS on it
Have to make the function parameters actually correct for what it needs, but otherwise it works fine enough. Should also move it into either the AI file or into the engine's file. The latter is a bit iffy because this is probably the last updates this engine is getting and thus it'd be better to just have it in my game's files instead.